requirements.txt
click>=8.1
rich>=13.0
streamlit>=1.30
pandas>=2.0
numpy>=1.24
plotly>=5.18
notion-client>=2.2
pyyaml>=6.0
